Refactor Scanner
I refactored the scanner to be my own component
based only on quagga2. I also changed it to a single
location, instead of locating for hopefully better performance.
It is also a fixed width so there aren't parts of the
fov moving off the screen.
I have also set multiple to false and frequency very high so that
it should resolve the correct barcode faster